+++ polyfuzz/models/__init__.py
@@ -11,10 +11,35 @@ try:
 except ModuleNotFoundError as e:
     Embeddings = NotInstalled("Flair and Huggingface Transformer Models", "flair")
 
+try:
+    from._sbert import SentenceEmbeddings
+except ModuleNotFoundError as e:
+    SentenceEmbeddings = NotInstalled("SentenceTransformers", "sbert")
+
+try:
+    from._gensim import GensimEmbeddings
+except ModuleNotFoundError as e:
+    GensimEmbeddings = NotInstalled("Gensim", "gensim")
+
+try:
+    from._spacy import SpacyEmbeddings
+except ModuleNotFoundError as e:
+    SpacyEmbeddings = NotInstalled("Spacy", "spacy")
+
+try:
+    from._use import USEEmbeddings
+except ModuleNotFoundError as e:
+    USEEmbeddings = NotInstalled("USE", "use")
+
+
 __all__ = [
     "BaseMatcher",
     "EditDistance",
     "Embeddings",
+    "SentenceEmbeddings",
+    "GensimEmbeddings",
+    "SpacyEmbeddings",
+    "USEEmbeddings",
     "RapidFuzz",
     "TFIDF",
     "cosine_similarity"