+++ polyfuzz/models/_distance.py
@@ -45,7 +45,8 @@ class EditDistance(BaseMatcher):
 
     def match(self,
               from_list: List[str],
-              to_list: List[str] = None) -> pd.DataFrame:
+              to_list: List[str] = None,
+              **kwargs) -> pd.DataFrame:
         """ Calculate the edit distances between two list of strings
         by parallelizing the calculation and passing the lists in
         batches.