+++ polyfuzz/models/_spacy.py
@@ -1,24 +1,19 @@
-import warnings
-warnings.simplefilter(action='ignore', category=FutureWarning)
-
 import numpy as np
 import pandas as pd
-from typing import List, Union
-from sklearn.preprocessing import normalize
-from flair.embeddings import DocumentPoolEmbeddings, WordEmbeddings, TokenEmbeddings
-from flair.data import Sentence
+from typing import List
+import spacy
 
 from ._utils import cosine_similarity
 from ._base import BaseMatcher
 
 
-class Embeddings(BaseMatcher):
+class SpacyEmbeddings(BaseMatcher):
     """
     Embed words into vectors and use cosine similarity to find
     the best matches between two lists of strings
 
     Arguments:
-        embedding_method: list of Flair embeddings to use
+        embedding_model: The Spacy model to use, this can be either a string or the model directly
         min_similarity: The minimum similarity between strings, otherwise return 0 similarity
         top_n: The number of best matches you want returned
         cosine_method: The method/package for calculating the cosine similarity.
@@ -34,31 +29,19 @@ class Embeddings(BaseMatcher):
     Usage:
 
     ```python
-    model = Embeddings(min_similarity=0.5)
+    distance_model = SpacyEmbeddings("en_core_web_md", min_similarity=0.5)
     ```
 
-    Or if you want a custom model to be used and it is a word embedding model,
-    pass it in as a list:
+    Or if you want to directly pass a Spacy model:
 
     ```python
-    embedding_model = WordEmbeddings('news')
-    model = Embeddings([embeddings_model], min_similarity=0.5)
-    ```
-
-    As you might have guessed, you can pass along multiple word embedding models and the
-    results will be averaged:
-
-    ```python
-    fasttext_embedding = WordEmbeddings('news')
-    glove_embedding = WordEmbeddings('glove')
-    bert_embedding = TransformerWordEmbeddings('bert-base-multilingual-cased')
-    model = Embeddings([glove_embedding,
-                        fasttext_embedding,
-                        bert_embedding ], min_similarity=0.5)
+    import spacy
+    embedding_model = spacy.load("en_core_web_md", exclude=['tagger', 'parser', 'ner', 'attribute_ruler', 'lemmatizer'])
+    distance_model = SpacyEmbeddings(embedding_model, min_similarity=0.5)
     ```
     """
     def __init__(self,
-                 embedding_method: Union[List, None] = None,
+                 embedding_model = "en_core_web_md",
                  min_similarity: float = 0.75,
                  top_n: int = 1,
                  cosine_method: str = "sparse",
@@ -66,17 +49,13 @@ class Embeddings(BaseMatcher):
         super().__init__(model_id)
         self.type = "Embeddings"
 
-        if not embedding_method:
-            self.document_embeddings = DocumentPoolEmbeddings([WordEmbeddings('news')])
-
-        elif isinstance(embedding_method, list):
-            self.document_embeddings = DocumentPoolEmbeddings(embedding_method)
-
-        elif isinstance(embedding_method, TokenEmbeddings):
-            self.document_embeddings = DocumentPoolEmbeddings([embedding_method])
-
+        if isinstance(embedding_model, str):
+            self.embedding_model = spacy.load(embedding_model, exclude=['tagger', 'parser', 'ner', 'attribute_ruler', 'lemmatizer'])
+        elif "spacy" in str(type(embedding_model)):
+            self.embedding_model = embedding_model
         else:
-            self.document_embeddings = embedding_method
+            raise ValueError("Please select a correct Spacy model by either using a string such as 'en_core_web_md' "
+                             "or create a nlp model using: `nlp = spacy.load('en_core_web_md')")
 
         self.min_similarity = min_similarity
         self.top_n = top_n
@@ -136,10 +115,26 @@ class Embeddings(BaseMatcher):
 
     def _embed(self, strings: List[str]) -> np.ndarray:
         """ Create embeddings from a list of strings """
-        embeddings = []
-        for name in strings:
-            sentence = Sentence(name)
-            self.document_embeddings.embed(sentence)
-            embeddings.append(sentence.embedding.cpu().numpy())
+        # Extract embeddings from a transformer model
+        if "transformer" in self.embedding_model.component_names:
+            embeddings = []
+            for doc in strings:
+                try:
+                    embedding = self.embedding_model(doc)._.trf_data.tensors[-1][0].tolist()
+                except:
+                    embedding = self.embedding_model("An empty document")._.trf_data.tensors[-1][0].tolist()
+                embeddings.append(embedding)
+            embeddings = np.array(embeddings)
 
-        return np.array(normalize(embeddings), dtype="double")
+        # Extract embeddings from a general spacy model
+        else:
+            embeddings = []
+            for doc in strings:
+                try:
+                    vector = self.embedding_model(doc).vector
+                except ValueError:
+                    vector = self.embedding_model("An empty document").vector
+                embeddings.append(vector)
+            embeddings = np.array(embeddings)
+
+        return embeddings