+++ tests/test_polyfuzz.py
@@ -4,28 +4,40 @@ import pandas as pd
 from rapidfuzz import fuzz
 
 from polyfuzz import PolyFuzz
-from polyfuzz.models import EditDistance, TFIDF, RapidFuzz, BaseMatcher
+from polyfuzz.models import EditDistance, TFIDF, RapidFuzz, BaseMatcher, SentenceEmbeddings
 
 from tests.utils import get_test_strings
 
 from_list, to_list = get_test_strings()
+sentence_model = SentenceEmbeddings("all-MiniLM-L6-v2")
+
+
+import numpy as np
+import pandas as pd
+from rapidfuzz import fuzz
+
+from polyfuzz import PolyFuzz
+from polyfuzz.models import BaseMatcher
 
 
 class MyModel(BaseMatcher):
-    def match(self, from_list, to_list):
+    def match(self, from_list, to_list, **kwargs):
         # Calculate distances
-        matches = [[fuzz.ratio(from_string, to_string) / 100 for to_string in to_list] for from_string in from_list]
-
+        matches = [[fuzz.ratio(from_string, to_string) / 100 
+                   for to_string in to_list] for from_string in from_list]
+        
         # Get best matches
         mappings = [to_list[index] for index in np.argmax(matches, axis=1)]
         scores = np.max(matches, axis=1)
-
+        
         # Prepare dataframe
-        matches = pd.DataFrame({'From': from_list, 'To': mappings, 'Similarity': scores})
+        matches = pd.DataFrame({'From': from_list,
+                                'To': mappings, 
+                                'Similarity': scores})
         return matches
 
 
-@pytest.mark.parametrize("method", ["EditDistance", "TF-IDF"])
+@pytest.mark.parametrize("method", ["EditDistance", "TF-IDF", sentence_model, MyModel()])
 def test_base_model(method):
     model = PolyFuzz(method).match(from_list, to_list)
     matches = model.get_matches()
@@ -36,7 +48,30 @@ def test_base_model(method):
     assert list(matches.columns) == ['From', 'To', 'Similarity']
 
 
-@pytest.mark.parametrize("method", ["EditDistance", "TF-IDF"])
+@pytest.mark.parametrize("method", ["EditDistance", "TF-IDF", sentence_model, MyModel()])
+def test_fit_model(method):
+    model = PolyFuzz(method).fit(from_list, to_list)
+    matches = model.get_matches()
+
+    assert isinstance(matches, pd.DataFrame)
+    assert matches.Similarity.mean() > 0.3
+    assert len(matches) == 6
+    assert list(matches.columns) == ['From', 'To', 'Similarity']
+
+    results = model.transform(to_list)
+
+    if method == "TF-IDF":
+        key = "TF-IDF"
+    elif method == "EditDistance":
+        key = "EditDistance"
+    else:
+        key = list(results.keys())[0]
+    
+    assert isinstance(results[key], pd.DataFrame)
+    assert results[key].Similarity.sum() > 0
+
+
+@pytest.mark.parametrize("method", ["EditDistance", "TF-IDF", sentence_model, MyModel()])
 def test_grouper(method):
     model = PolyFuzz(method).match(from_list, to_list)
     model.group(link_min_similarity=0.75)